moved reducer out of file into own dir
